Fixed Typo for PR #118.
Also moved the URI_STUB param to all URIs.

This will need to be handled diffferntly for the full fix, and this workaround will need to be backed out.

The full solution will be to implement the 1.5 endpoints directly, and analyze and modify the appropriate methods to sue the 1.5 endpoints and needed.

Currently it looks like only multi-cloud endpoints use 1.5.